requirements.txt
# Content Creator System - Python Dependencies
# Last Updated: 2025-11-01

# ===== Core Framework =====
streamlit>=1.30.0
python-dotenv>=1.0.0

# ===== Notion Integration =====
notion-client>=2.2.0

# ===== OpenRouter / LLM Integration =====
# OpenAI SDK for OpenRouter compatibility
openai>=1.0.0

# ===== Background Job Scheduling =====
apscheduler>=3.10.0

# ===== Testing (TDD) =====
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ===== Utilities =====
requests>=2.31.0
pyyaml>=6.0
pydantic>=2.5.0

# ===== Optional: Social Media Publishing =====
# linkedin-api>=2.0.0  # Uncomment when implementing LinkedIn publishing
# facebook-sdk>=3.1.0  # Uncomment when implementing Facebook publishing

# ===== Development =====
black>=23.12.0  # Code formatting
flake8>=7.0.0   # Linting
